refactor: extract common controller logic to a new file
